4.tools_agents/requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
beautifulsoup4
pymupdf
arxiv
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain-huggingface
tqdm
faiss-cpu
langchain-chroma
streamlit
duckdb
langchain-groq
duckduckgo-search
SQLAlchemy
unstructured
youtube-transcript-api
validators
numexpr
pytube
openai